+++ Shared/Data/Dto/Setting/SettingsLogoDto.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace Data.Dto
 {
-    public  class SettingsLogoDto : BaseEntityDto<byte>
+    public  class SettingsLogoDto : BaseEntityDto<int>
     {
         #region Logo
